requirements.txt
streamlit==1.32.0
langchain-groq==0.1.9
python-dotenv==1.0.1
reportlab==4.1.0
Pillow==10.2.0
